requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Firebase dependencies
firebase-admin==6.2.0
requests==2.31.0

# AI and ML dependencies
google-generativeai==0.3.2
google-genai

# File processing dependencies
PyPDF2==3.0.1
openpyxl==3.1.2
python-docx==0.8.11
Pillow==10.0.1
pytesseract==0.3.10

# Environment and configuration
python-dotenv==1.0.0

# Email validation
email-validator==2.0.0

# Logging and monitoring
structlog==23.1.0

# Security
cryptography==41.0.4

# Development and testing
pytest==7.4.0
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0

# Production server
gunicorn==21.2.0
